Chỉnh sang thymeleaf

+++ UTESocial/src/main/java/aloute/com/repository/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import aloute.com.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
-	User findByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    User findByEmail(String email);
 }
